feat: Move isAbsoluteUri to utils
