+++ src/api/Api.js
@@ -2,7 +2,7 @@ import axios from 'axios'
 import { stringify } from 'qs'
 import { validateCallbackFn } from '../helpers/validateCallbackFn'
 import { setCallbackFns } from '../helpers/setCallbackFns'
-import { isAbsoluteUri } from '../helpers/isAbsoluteUri'
+import { isAbsoluteUri } from '../shared/utils'
 
 /**
  * Wrapper around json:api requests, sets content type and other defaults.
