+++ src/module/ModuleBuilder.js
@@ -4,7 +4,6 @@ import { checkConfigProperty } from '../helpers/checkConfigProperty'
 import { deleteAction } from './actions/deleteAction'
 import { getAction } from './actions/getAction'
 import { getProperty } from './getters/getProperty'
-import { isAbsoluteUri } from '../helpers/isAbsoluteUri'
 import { listAction } from './actions/listAction'
 import { listRelatedAction } from './actions/listRelatedAction'
 import { itemsInRelationshipFormat } from './getters/itemsInRelationshipFormat'
@@ -21,7 +20,7 @@ import { startLoadingMutation, endLoadingMutation } from './mutations/loading'
 import { updateMutation } from './mutations/updateMutation'
 import { hasChanges } from './getters/hasChanges'
 import { ResourceBuilder } from './resource/ResourceBuilder'
-import { hasOwn } from '../shared/utils'
+import { hasOwn, isAbsoluteUri } from '../shared/utils'
 import { Performance } from '../shared/Performance'
 
 /**
